Fix type of pc

+++ src/rv32i/rv32i.go
@@ -49,8 +49,8 @@ func Srai(inst instruction.Instruction, cpu *cpu.Cpu) {
 }
 
 func Jal(inst instruction.Instruction, cpu *cpu.Cpu) {
-	cpu.Registers[inst.Rd] = int(cpu.Pc) + 4
-	cpu.Pc += uint32(inst.Imm)
+	cpu.Registers[inst.Rd] = cpu.Pc + 4
+	cpu.Pc += inst.Imm
 }
